Standalone executable seems to work allright

+++ src/cavity/GePolCavity.cpp
@@ -158,7 +158,7 @@ void GePolCavity::build(int maxts, int maxsph, int maxvert)
 
     double * rin = radii_scratch.data();
     double * mass = new double[molecule_.nAtoms()];
-    for (int i = 0; i < molecule_.nAtoms(); ++i) {
+    for (size_t i = 0; i < molecule_.nAtoms(); ++i) {
 	    mass[i] = molecule_.masses(i);
     }
